\name{FMT-package}
\alias{FMT-package}
\alias{FMT}
\docType{package}
\title{
   open source Forest Management library
}
\description{
  A more detailed description of what the package does. A length
  of about one to five lines is recommended.
}
\details{
  FMT is a C++ library made to interpret forest planning models based on Woodstock files formulation. FMT can be used through Python, Cran-R or directly from C++. FMT's generic approach to forest planning gives the opportunity to the user to manipulate forest planning models and generate forest planning solutions. The objective behind this project is to bring all forest planning models types onto the same generic library to help the community to develop new approaches to forest planning problems or improve the one that have been there for years.
}
\author{
Guillaume Cyr

Maintainer: Guillaume Cyr <Guillaume.Cyr@fec.gouv.qc.ca>
}
\references{
 
}
\keyword{ package }
\seealso{
  https://github.com/gcyr/FMT.git
}
\examples{
  \dontrun{
     
  }
}
