% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{plotFMT}
\alias{plotFMT}
\title{Plot fitting results from fmt function}
\usage{
plotFMT(x, type, ...)
}
\arguments{
\item{x}{return list from fmt function}

\item{type}{type of plots}

\item{...}{extra parameters for plot function}
}
\value{
No return value.
}
\description{
This function provides fitting plots of fmt function.
}
